+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## [Unreleased]
+### Changed
+- Checks for clone to relative directory.
+
 ## 1.0.1 - 2017-07-03
 ### Changed
 - Updated README.